requirements.txt
asgiref==3.2.10
certifi==2020.12.5
chardet==4.0.0
Django==3.0.8
django-cors-headers==3.6.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2020.1
requests==2.25.1
sqlparse==0.3.1
urllib3==1.26.2
